+++ spec/Badger/Bundle/StorageUtilsBundle/Doctrine/Remover/BaseRemoverSpec.php
@@ -1,32 +1,31 @@
 <?php
 
-namespace spec\Badger\StorageUtilsBundle\Doctrine\Saver;
+namespace spec\Badger\Bundle\StorageUtilsBundle\Doctrine\Remover;
 
 use Badger\Component\Game\Model\BadgeInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 use PhpSpec\ObjectBehavior;
 
-class BaseSaverSpec extends ObjectBehavior
+class BaseRemoverSpec extends ObjectBehavior
 {
     function let(ObjectManager $objectManager)
     {
         $this->beConstructedWith($objectManager, 'Badger\Component\Game\Model\BadgeInterface');
     }
 
-    function it_is_a_saver()
+    function it_is_a_remover()
     {
-        $this->shouldHaveType('Badger\Component\StorageUtils\Saver\SaverInterface');
+        $this->shouldHaveType('Badger\Component\StorageUtils\Remover\RemoverInterface');
     }
 
-    function it_persists_the_object_and_flushes_the_unit_of_work($objectManager, BadgeInterface $type)
+    function it_removes_the_object_and_flushes_the_unit_of_work($objectManager, BadgeInterface $type)
     {
-        $objectManager->persist($type)->shouldBeCalled();
+        $objectManager->remove($type)->shouldBeCalled();
         $objectManager->flush()->shouldBeCalled();
-
-        $this->save($type);
+        $this->remove($type);
     }
 
-    function it_throws_exception_when_save_anything_else_than_the_expected_class()
+    function it_throws_exception_when_remove_anything_else_than_the_expected_class()
     {
         $anythingElse = new \stdClass();
         $exception = new \InvalidArgumentException(
@@ -35,6 +34,6 @@ class BaseSaverSpec extends ObjectBehavior
                 get_class($anythingElse)
             )
         );
-        $this->shouldThrow($exception)->during('save', [$anythingElse]);
+        $this->shouldThrow($exception)->during('remove', [$anythingElse]);
     }
 }
