+++ spec/Badger/Bundle/StorageUtilsBundle/Doctrine/Saver/BaseSaverSpec.php
@@ -1,31 +1,32 @@
 <?php
 
-namespace spec\Badger\StorageUtilsBundle\Doctrine\Remover;
+namespace spec\Badger\Bundle\StorageUtilsBundle\Doctrine\Saver;
 
 use Badger\Component\Game\Model\BadgeInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 use PhpSpec\ObjectBehavior;
 
-class BaseRemoverSpec extends ObjectBehavior
+class BaseSaverSpec extends ObjectBehavior
 {
     function let(ObjectManager $objectManager)
     {
         $this->beConstructedWith($objectManager, 'Badger\Component\Game\Model\BadgeInterface');
     }
 
-    function it_is_a_remover()
+    function it_is_a_saver()
     {
-        $this->shouldHaveType('Badger\Component\StorageUtils\Remover\RemoverInterface');
+        $this->shouldHaveType('Badger\Component\StorageUtils\Saver\SaverInterface');
     }
 
-    function it_removes_the_object_and_flushes_the_unit_of_work($objectManager, BadgeInterface $type)
+    function it_persists_the_object_and_flushes_the_unit_of_work($objectManager, BadgeInterface $type)
     {
-        $objectManager->remove($type)->shouldBeCalled();
+        $objectManager->persist($type)->shouldBeCalled();
         $objectManager->flush()->shouldBeCalled();
-        $this->remove($type);
+
+        $this->save($type);
     }
 
-    function it_throws_exception_when_remove_anything_else_than_the_expected_class()
+    function it_throws_exception_when_save_anything_else_than_the_expected_class()
     {
         $anythingElse = new \stdClass();
         $exception = new \InvalidArgumentException(
@@ -34,6 +35,6 @@ class BaseRemoverSpec extends ObjectBehavior
                 get_class($anythingElse)
             )
         );
-        $this->shouldThrow($exception)->during('remove', [$anythingElse]);
+        $this->shouldThrow($exception)->during('save', [$anythingElse]);
     }
 }
