+++ spec/Badger/Bundle/TagBundle/Security/TagVoterSpec.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace spec\Badger\TagBundle\Security;
+namespace spec\Badger\Bundle\TagBundle\Security;
 
 use Badger\Component\Game\Model\BadgeInterface;
 use Badger\Component\Tag\Model\TagInterface;
-use Badger\TagBundle\Security\TagVoter;
+use Badger\Bundle\TagBundle\Security\TagVoter;
 use Badger\Component\Tag\Taggable\TaggableInterface;
 use Badger\Component\User\Model\UserInterface;
 use Doctrine\Common\Collections\ArrayCollection;
